feat: buildkit을 이용한 ecr docker cache 기능 추가
- Buildkit 사용해도
RUN apt-get install~ 은 cache 되는 데
Pip install은 cache안된다.
Multi stage build 등 필요할지도

+++ src/solution_register.py
@@ -1114,8 +1114,7 @@ class SolutionRegister:
         ecr_scope = self.infra_setup["WORKSPACE_NAME"].split('-')[0] # magna-ws --> magna
         self.ecr_repo = self.ecr_name.split("/")[1] + '/' + ecr_scope + "/ai-solutions/" + self.solution_name + "/" + self.pipeline + "/"  + self.solution_name  
         self.ecr_full_url = self.ecr_url + '/' + self.ecr_repo 
-
-        ## 동일 이름의 ECR 존재 시, 삭제하고 다시 생성한다. 
+        # get ecr client 
         try: 
             try:
                 ecr_client = self.session.client('ecr',region_name=self.infra_setup['REGION'])
@@ -1124,13 +1123,29 @@ class SolutionRegister:
                 ecr_client = boto3.client('ecr', region_name=self.infra_setup['REGION'])
         except Exception as e:
             raise ValueError(f"Failed to create ecr client. \n {str(e)}")
-
-        try:
-            ecr_client.delete_repository(repositoryName=self.ecr_repo, force=True)
-            print_color(f"[SYSTEM] Repository {self.ecr_repo} already exists. Deleting...", color='yellow')
-        except Exception as e:
-            print_color(f"[WARNING] Failed to delete pre-existing ECR Repository. \n {str(e)}", color='yellow')
-
+        ## 동일 이름의 ECR 존재 시, 삭제하고 다시 생성한다. 
+        ## 240324 solution update 시에는 본인 버전만 삭제해야지 통째로 repo 삭제하면 cache 기능 사용불가 
+        if self.solution_info['solution_update'] == False:
+            try:
+                ecr_client.delete_repository(repositoryName=self.ecr_repo, force=True)
+                print_color(f"[SYSTEM] Repository {self.ecr_repo} already exists. Deleting...", color='yellow')
+            except Exception as e:
+                print_color(f"[WARNING] Failed to delete pre-existing ECR Repository. \n {str(e)}", color='yellow')
+        else: 
+            try:
+                print_color(f"Now in solution update mode. Only delete current version docker image.", color='yellow')
+                resp_ecr_image_list = ecr_client.list_images(repositoryName=self.ecr_repo)
+                print(resp_ecr_image_list)
+                cur_ver_image = []
+                for image in resp_ecr_image_list['imageIds']:
+                    if 'imageTag' in image.keys():
+                        if image['imageTag'] == f'v{self.solution_version_new}':
+                            cur_ver_image.append(image)
+                # 사실 솔루션 업데이트 시엔 이미 만들어진 현재 버전 이미지가 거의 없을 것임
+                if len(cur_ver_image) != 0: 
+                    resp_delete_cur_ver = ecr_client.batch_delete_image(repositoryName=self.ecr_repo, imageIds=cur_ver_image)
+            except Exception as e:
+                raise NotImplementedError(f'Failed to delete current versioned image \n {str(e)}') 
         print_color(f"[SYSTEM] target AWS ECR url: \n", color='blue')
         print(f"{self.ecr_url}")
         try:
@@ -1181,24 +1196,26 @@ class SolutionRegister:
             "--repository-name", self.ecr_repo,
             "--image-scanning-configuration", "scanOnPush=true",
             ]
-        # create ecr repo
-        try:
-            create_resp = ecr_client.create_repository(repositoryName=self.ecr_repo)
-            repository_arn = create_resp['repository']['repositoryArn']
-            tags_new = []
-            for tag in tags:
-                    key, value = tag.split(',')
-                    tag_dict = {'Key': key.split('=')[1], 'Value': value.split('=')[1]}
-                    tags_new.append(tag_dict)
+        # create ecr repo 
+        # 240324 solution update True일 땐 이미지만 삭제했으므로 repo를 재생성하진 않는다. 
+        if self.solution_info['solution_update'] == False:
+            try:
+                create_resp = ecr_client.create_repository(repositoryName=self.ecr_repo)
+                repository_arn = create_resp['repository']['repositoryArn']
+                tags_new = []
+                for tag in tags:
+                        key, value = tag.split(',')
+                        tag_dict = {'Key': key.split('=')[1], 'Value': value.split('=')[1]}
+                        tags_new.append(tag_dict)
 
-            resp = ecr_client.tag_resource(
-                resourceArn=repository_arn,
-                tags=tags_new
-                )
-            print_color(f"[SYSTEM] AWS ECR create-repository response: ", color='cyan')
-            print(f"{resp}")
-        except Exception as e:
-            raise NotImplementedError(f"Failed to AWS ECR create-repository:\n + {e}")
+                resp = ecr_client.tag_resource(
+                    resourceArn=repository_arn,
+                    tags=tags_new
+                    )
+                print_color(f"[SYSTEM] AWS ECR create-repository response: ", color='cyan')
+                print(f"{resp}")
+            except Exception as e:
+                raise NotImplementedError(f"Failed to AWS ECR create-repository:\n + {e}")
 
     def _aws_codebuild(self):
         # 0. create boto3 session and get codebuild service role arn 
@@ -1318,7 +1335,13 @@ class SolutionRegister:
             ## {'version': 0.2, 'phases': {'pre_build': {'commands': None}, 'build': {'commands': None}, 'post_build': {'commands': None}}}
             buildspec = yaml.safe_load(file)
         pre_command = [f'aws ecr get-login-password --region {self.infra_setup["REGION"]} | docker login --username AWS --password-stdin {self.ecr_url}/{self.ecr_repo}']
-        build_command = [f'docker build . -t {self.ecr_full_url}:v{self.solution_version_new}']
+        build_command = ['export DOCKER_BUILDKIT=1']
+        if self.solution_info['solution_update'] == True: 
+            # 이전 version docker 다운로드 후 이번 version build 시 cache 활용
+            pre_command.append(f'docker pull {self.ecr_full_url}:v{self.solution_version_new - 1}')
+            build_command.append(f'docker build --build-arg BUILDKIT_INLINE_CACHE=1 --cache-from {self.ecr_full_url}:v{self.solution_version_new - 1} -t {self.ecr_full_url}:v{self.solution_version_new} .')
+        else:
+            build_command.append(f'docker build --build-arg BUILDKIT_INLINE_CACHE=1 -t {self.ecr_full_url}:v{self.solution_version_new} .')
         post_command = [f'docker push {self.ecr_full_url}:v{self.solution_version_new}']
         buildspec['phases']['pre_build']['commands'] = pre_command
         buildspec['phases']['build']['commands'] = build_command
@@ -1337,11 +1360,17 @@ class SolutionRegister:
                 'mkdir -p ~/.docker/cli-plugins', \
                 'mv buildx-v0.4.2.linux-amd64 ~/.docker/cli-plugins/docker-buildx', \
                 'chmod a+rx ~/.docker/cli-plugins/docker-buildx', \
-                'docker run --rm tonistiigi/binfmt --install all']
-                #'docker run --privileged --rm tonistiigi/binfmt --install all']
+                #'docker run --rm tonistiigi/binfmt --install all']
+                'docker run --privileged --rm tonistiigi/binfmt --install all']
         pre_command = [f'aws ecr get-login-password --region {self.infra_setup["REGION"]} | docker login --username AWS --password-stdin {self.ecr_url}/{self.ecr_repo}']
-        build_command = ['docker buildx create --use --name crossx', \
-                f'docker buildx build --push --platform=linux/amd64,linux/arm64 -t {self.ecr_full_url}:v{self.solution_version_new} .']
+        build_command = ['export DOCKER_BUILDKIT=1', \
+                    'docker buildx create --use --name crossx']
+        if self.solution_info['solution_update'] == True: 
+            # 이전 version docker 다운로드 후 이번 version build 시 cache 활용 
+            pre_command.append(f'docker pull {self.ecr_full_url}:v{self.solution_version_new - 1}')
+            build_command.append(f'docker buildx build --push --platform=linux/amd64,linux/arm64 --build-arg BUILDKIT_INLINE_CACHE=1 --cache-from {self.ecr_full_url}:v{self.solution_version_new - 1} -t {self.ecr_full_url}:v{self.solution_version_new} .')
+        else: 
+            build_command.append(f'docker buildx build --push --platform=linux/amd64,linux/arm64 --build-arg BUILDKIT_INLINE_CACHE=1 -t {self.ecr_full_url}:v{self.solution_version_new} .')
         # buildspec['phases']['install']['runtime-versions'] = runtime_docker_version
         buildspec['phases']['install']['commands'] = install_command
         buildspec['phases']['pre_build']['commands'] = pre_command
@@ -1403,7 +1432,7 @@ class SolutionRegister:
             data = {'container_uri': f'{self.ecr_full_url}:v{self.solution_version_new}'} # full url 는 tag 정보까지 포함 
             self.sm_yaml['pipeline'][self.sm_pipe_pointer].update(data)
             print_color(f"[SYSTEM] Completes setting << container_uri >> in solution_metadata.yaml:", color='green')
-            print(f"pipeline: -container_uri: {data['container_uri']}")
+            print(f"container_uri: {data['container_uri']}")
             self._save_yaml()
         except Exception as e: 
             raise NotImplementedError(f"Failed to set << container_uri >> in the solution_metadata.yaml \n {str(e)}")
@@ -1567,7 +1596,8 @@ class SolutionRegister:
             path = REGISTER_INTERFACE_PATH + self.solution_file
             msg = f"[SYSTEM] AI solution 등록 정보를 {path} 에서 확인합니다."
             load_response = self._load_response_yaml(path, msg)
-
+        print_color('load_response: \n', color='blue')
+        print(load_response)
         ########################
         ### instance name      -- spec 변경 시, 수정 필요 (fix date: 24.02.23)
         name = load_response['name'] + \
